Finish role specifications. continue to student dashboard.
