+++ src/Components/Dashboards/AdminDashboard/index.js
@@ -1,0 +1,1 @@
+export { default } from './AdminDashboard';
