+++ src/Components/Dashboards/FacultyDashboard/FacultyDashboard.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function FacultyDashboard() {
+  return (
+    <div>FacultyDashboard</div>
+  )
+}
+
+export default FacultyDashboard
