+++ src/Components/Dashboards/StudentDashboard/StudentDashboard.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function StudentDashboard() {
+  return (
+    <div>StudentDashboard</div>
+  )
+}
+
+export default StudentDashboard
