+++ src/Pages/DashboardPage/DashboardPage.css
@@ -1,296 +1,22 @@
-/* General Stylings */
-.Dashboard-PageContentContainer {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-}
-
-.Dashboard-PageHeaderContainer {
+.Dashboard-MessageContainer {
+    height: 100dvh;
     width: 100%;
     display: flex;
     flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    box-sizing: border-box;
-}
-
-.Dashboard-ContentContainer {
-    width: 100%;
-    flex: 1;
-    display: flex;
-    flex-direction: row;
-    overflow-y: auto;
-}
-
-.Dashboard-LeftContentPanel {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    box-sizing: border-box;
-    overflow-y: hidden;
-}
-
-.Dashboard-SectionHeader {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
+    justify-content: center;
     align-items: center;
 }
 
-.Dashboard-InventorySection {
-    width: 100%;
+.Dashboard-Message {
     display: flex;
     flex-direction: column;
-    justify-content: flex-start;
-}
-
-.Dashboard-InventorySummaryList {
-    width: 100%;
-}
-
-.Dashboard-ReservationSection {
-    width: 100%;
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    overflow-y: auto;
-}
-
-.Dashboard-ReservationFilterContainer {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-}
-
-.Dashboard-ReservationList {
-    width: 100%;
-}
-
-.Dashboard-RightContent {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    box-sizing: border-box;
-}
-
-.Dashboard-InUseSection {
-    width: 100%;
-}
-
-.Dashboard-UnderRepairSection,
-.Dashboard-ReservationDetailSection {
-    width: 100%;
-    overflow-y: auto;
-}
-
-.Dashboard-ReservationDetailSection {
-    flex: 1;
-}
-
-.Dashboard-ReservationActionContainer {
-    display: flex;
-    flex-direction: row;
     align-items: center;
+    justify-content: center;
 }
 
-.Dashboard-ReservationActionButton {
-    width: 100%;
-}
-
-.Dashboard-ReservationActionButton:first-child {
-    margin-left: 0px;
-}
-
-.Dashboard-Hide {
-    display: none;
-}
-
-/* For tablets and any larger devices */
-@media screen and (min-width: 481px) {
-    .Dashboard-PageContentContainer {
-        height: 100%;
-    }
-
-    .Dashboard-PageHeaderContainer {
-        height: calc(9.45625vmin);
-        border-bottom: calc(0.278125vmin) solid var(--Gray1);
-        padding: calc(2.225vmin) calc(2.225vmin);
-    }
-
-    .Dashboard-LogoContainer {
-        display: none;
-    }
-
-    .Dashboard-ActionContainer {
-        display: none;
-    }
-
-    .Dashboard-ContentContainer {
-        height: 100% - calc(9.45625vmin);
-    }
-
-    .Dashboard-LeftContentPanel {
-        flex: 1;
-        height: calc(100% - calc(4.45vmin));
-        margin: calc(2.225vmin) calc(2.225vmin);
-    }
-
-    .Dashboard-SectionHeader {
-        height: calc(calc(6.675vmin) + calc(2 * calc(0.278125vmin)));
-    }
-
-    .Dashboard-InventorySection {
-        height: calc(4 * calc(2.225vmin) + calc(calc(34.4875vmin) + calc(1vmin)));
-        margin-bottom: calc(4.171875vmin)
-    }
-
-    .Dashboard-SearchBar {
-        width: calc(38.9375vmin);
-    }
-
-    .Dashboard-InventorySummaryList {
-        height: calc(calc(100%) - calc(6.675vmin));
-        margin-top: calc(2.225vmin);
-    }
-
-    .Dashboard-ReservationList {
-        height: calc(calc(100%) - calc(6.675vmin));
-        margin-top: calc(2.225vmin);
-    }
-
-    .Dashboard-RightContentPanel {
-        width: calc(calc(45.6125vmin));
-        height: 100%;
-        border-left: calc(0.278125vmin) solid var(--Gray1);
-    }
-
-    .Dashboard-RightContent {
-        height: calc(100% - calc(2.225vmin));
-        padding: calc(2.25vmin) calc(2.225vmin) 0px calc(2.225vmin);
-    }
-
-    .Dashboard-AddEquipmentButton {
-        width: 100%;
-    }
-
-    .Dashboard-InUseSection {
-        height: calc(4 * calc(2.225vmin) + calc(calc(34.4875vmin) + calc(1vmin)));
-        margin-bottom: calc(4.171875vmin);
-    }
-    
-    .Dashboard-ReservationActionContainer {
-        margin-top: calc(2.225vmin);
-    }
-
-    .Dashboard-ReservationActionButton {
-        margin-left: calc(2.225vmin);
-    }
-}
-
-/* For mobile and any smaller devices */
-@media screen and (max-width: 480px) {
-    .Dashboard-PageContentContainer {
-        height: 0px;
-    }
-
-    .Dashboard-PageHeaderContainer {
-        height: calc(17.8vmin);
-        border-bottom: calc(0.55625vmin) solid var(--Gray1);
-        padding: calc(4.45vmin) calc(4.45vmin);
-    }
-
-    .Dashboard-PageHeaderContainer p.heading-2 {
-        font-size: calc(5.562vmin);
-        margin-left: calc(4.45vmin);
-    }
-    
-    .Dashboard-PageHeader {
-        display: flex;
-        flex-direction: row;
-        justify-content: flex-start;
-    }
-
-    .Dashboard-LogoContainer {
-        width: calc(8.9vmin); 
-    }
-    
-    .Dashboard-LogoContainer img {
-        width:100%;
-        height:100%;
-    }
-
-    .Dashboard-ActionContainer {
-        width: calc(24.475vmin);
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-    }
-
-    .Dashboard-ActionButton {
-        width: calc(6.675vmin);
-    }
-
-    .Dashboard-ContentContainer {
-        height: 100% - calc(17.8vmin);
-    }
-    
-    .Dashboard-LeftContentPanel {
-        width: 100%;
-        height: calc(100% - calc(4.45vmin));
-        margin: calc(4.45vmin) calc(4.45vmin) 0px calc(4.45vmin);
-    }    
-
-    .Dashboard-SectionHeader {
-        height: calc(calc(13.35vmin) + calc(2 * calc(0.55625vmin)));
-    }
-
-    .Dashboard-InventorySection {
-        height: calc(67.8625vmin);
-        margin-bottom: calc(8.34375vmin);
-    }
-    
-    .Dashboard-SearchBar {
-        width: calc(64.525vmin);
-    }
-
-    .Dashboard-InventorySummaryList {
-        height: calc(calc(100%) - calc(13.35vmin));
-        margin-top: calc(4.45vmin);
-    }
-
-    .Dashboard-ReservationList {
-        height: calc(calc(100%) - calc(13.35vmin));
-        margin-top: calc(4.45vmin);
-    }
-
-    .Dashboard-RightContentPanel {
-        display: none;
-    }
-
-    .Dashboard-RightContentPanelActive {
-        width: 100%;
-        height: 100%;
-    }
-
-    .Dashboard-RightContent {
-        height: 100%;
-        padding: calc(4.45vmin) calc(4.45vmin) 0px calc(4.45vmin);
-    }
-
-    .Dashboard-InUseSection {
-        width: 100%;
-        height: calc(67.8625vmin);
-        margin-bottom: calc(8.34375vmin);
-    }
-
-    .Dashboard-ReservationActionContainer {
-        margin-bottom: calc(4.45vmin);
-    }
-
-    .Dashboard-ReservationActionButton {
-        margin-left: calc(4.45vmin);
-    }
+.Dashboard-MessageActionContainer {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
 }
